+++ app/Repository/Publisher/MySqlStatsRepository.php
@@ -802,7 +802,7 @@ class MySqlStatsRepository implements StatsRepository
         array $result,
         ChartResolution $resolution,
     ): array {
-        if (count($result) === 0) {
+        if (empty($result)) {
             return [];
         }
 
@@ -911,7 +911,7 @@ class MySqlStatsRepository implements StatsRepository
 
     private static function overwriteStartDate(DateTimeInterface $dateStart, array $result): array
     {
-        if (count($result) > 0) {
+        if (!empty($result)) {
             $result[0][0] = $dateStart->format(DateTimeInterface::ATOM);
         }