+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Command deleting users
+### Changed
+- Single line logs
+### Fixed
+- Joining fee does not have to be greater than required
 
 ## [2.8.0] - 2024-03-19
 ### Added