+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Command deleting users
 - Additional adserver statistics
+- Securing adserver statistics
 ### Changed
 - Single line logs
 ### Fixed