+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.9.0] - 2024-06-11
 ### Added
 - Command deleting users
 - Additional adserver statistics
@@ -988,7 +990,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v2.8.0...develop
+[Unreleased]: https://github.com/adshares/adserver/compare/v2.9.0...develop
+[2.9.0]: https://github.com/adshares/adserver/compare/v2.8.0...v2.9.0
 [2.8.0]: https://github.com/adshares/adserver/compare/v2.7.0...v2.8.0
 [2.7.0]: https://github.com/adshares/adserver/compare/v2.6.4...v2.7.0
 [2.6.4]: https://github.com/adshares/adserver/compare/v2.6.3...v2.6.4